% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padroniza_idade.R
\name{padroniza_idade}
\alias{padroniza_idade}
\title{Padroniza Idade}
\usage{
padroniza_idade(x)
}
\arguments{
\item{x}{Um data frame contendo dados a serem padronizados.}
}
\value{
Um data frame com idades padronizadas.
}
\description{
Esta função padroniza as idades no formato padrão do SIM em faixas etárias quinquenais.
}
\examples{
\dontrun{
dados <- data.frame(IDADE = c("001", "002", "090", NA))
dados_padronizados <- padroniza_idade(dados)
}
}
