% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redistribuicao_causas_externas.R
\name{redistribuicao_causas_externas}
\alias{redistribuicao_causas_externas}
\title{redistribuicao_causas_externas}
\usage{
redistribuicao_causas_externas(dados_completos, dados_redis)
}
\arguments{
\item{dados_completos, dados_redis, }{padrao vetor de números.}
}
\description{
Carregar as bases de dados do SIM.
}
