% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coletor_dados.R
\name{Coletor_dados}
\alias{Coletor_dados}
\title{Coletor_dados}
\usage{
Coletor_dados(caminho, padrao)
}
\arguments{
\item{caminho, }{padrao vetor de números.}
}
\description{
Carregar as bases de dados do SIM.
}
