% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padroniza_idade.R
\name{padroniza_idade}
\alias{padroniza_idade}
\title{padroniza_idade}
\usage{
padroniza_idade(x, caminho_arq_idade)
}
\arguments{
\item{x, caminho_arq_idade, }{padrao vetor de números.}
}
\description{
Padroniza a idade de acordo com as faixas etárias utilizadas no script.
}
