% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepara_base_generalizada.R
\name{prepara_base_generalizada}
\alias{prepara_base_generalizada}
\title{prepara_base_generalizada}
\usage{
prepara_base_generalizada(base_dados_completos)
}
\arguments{
\item{base_dados_completos}{padrao vetor de números.}
}
\description{
Carregar as bases de dados do SIM.
}
