% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redistribuicao_causas_ivestigacao.R
\name{redistribuicao_causas_ivestigacao}
\alias{redistribuicao_causas_ivestigacao}
\title{redistribuicao_causas_ivestigacao}
\usage{
redistribuicao_causas_ivestigacao(dados_completos, dados_redis, pesos)
}
\arguments{
\item{dados_completos, dados_redis, pesos, }{padrao vetor de números.}
}
\description{
Carregar as bases de dados do SIM.
}
