% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redistribuicao_causas_mat_inf.R
\name{redistribuicao_causas_mat_inf}
\alias{redistribuicao_causas_mat_inf}
\title{redistribuicao_causas_mat_inf}
\usage{
redistribuicao_causas_mat_inf(dados_completos, dados_redis)
}
\arguments{
\item{dados_completos, dados_redis, }{padrao vetor de números.}
}
\description{
Carregar as bases de dados do SIM.
}
