% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redistribuicao_dados_faltantes.R
\name{redistribuicao_dados_faltantes}
\alias{redistribuicao_dados_faltantes}
\title{redistribuicao_dados_faltantes}
\usage{
redistribuicao_dados_faltantes(base_prop, dados_ign)
}
\arguments{
\item{base_prop, dados_ign, }{padrao vetor de números.}
}
\description{
Carregar as bases de dados do SIM.
}
