% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separa_reg_GC.R
\name{separa_reg_GC}
\alias{separa_reg_GC}
\title{separa_reg_GC}
\usage{
separa_reg_GC(dados_sem_ign, dados_municipios)
}
\arguments{
\item{dados_sem_ign, dados_municipios, }{padrao vetor de números.}
}
\description{
Carregar as bases de dados do SIM.
}
