% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_final_1.R
\name{tabela_final_1}
\alias{tabela_final_1}
\title{tabela_final_1}
\usage{
tabela_final_1(x, arq_ICD_map, arq_regioes)
}
\arguments{
\item{x, arq_ICD_map, arq_regioes, }{padrao vetor de números.}
}
\description{
Carregar as bases de dados do SIM.
}
